Break up the api calls
